+++ src/test/kotlin/com/wutsi/codegen/kotlin/server/ServerCodeGeneratorTest.kt
@@ -38,7 +38,7 @@ internal class ServerCodeGeneratorTest {
     fun `generators`() {
         val codegen = ServerCodeGenerator(mapper = KotlinMapper(context))
 
-        assertEquals(19, codegen.generators.size)
+        assertEquals(16, codegen.generators.size)
         assertTrue(codegen.generators[0] is ServerDtoCodeGenerator)
         assertTrue(codegen.generators[1] is ServerDelegateCodeGenerator)
         assertTrue(codegen.generators[2] is ServerControllerCodeGenerator)
@@ -49,14 +49,11 @@ internal class ServerCodeGeneratorTest {
         assertTrue(codegen.generators[7] is EditorConfigCodeGenerator)
         assertTrue(codegen.generators[8] is ServerGithubActionsCodeGenerator)
         assertTrue(codegen.generators[9] is GitCodeGenerator)
-        assertTrue(codegen.generators[10] is CacheCodeGenerator)
-        assertTrue(codegen.generators[11] is MQueueCodeGenerator)
-        assertTrue(codegen.generators[12] is SwaggerCodeGenerator)
-        assertTrue(codegen.generators[13] is ServerReadmeCodeGenerator)
-        assertTrue(codegen.generators[14] is CorsCodeGenerator)
-        assertTrue(codegen.generators[15] is TracingCodeGenerator)
-        assertTrue(codegen.generators[16] is DatabaseCodeGenerator)
-        assertTrue(codegen.generators[17] is SecurityCodeGenerator)
-        assertTrue(codegen.generators[18] is RenovateCodeGenerator)
+        assertTrue(codegen.generators[10] is SwaggerCodeGenerator)
+        assertTrue(codegen.generators[11] is ServerReadmeCodeGenerator)
+        assertTrue(codegen.generators[12] is CorsCodeGenerator)
+        assertTrue(codegen.generators[13] is DatabaseCodeGenerator)
+        assertTrue(codegen.generators[14] is SecurityCodeGenerator)
+        assertTrue(codegen.generators[15] is RenovateCodeGenerator)
     }
 }